Add restore functionality to trash Use `StorageList` for usage Tweak styles

+++ storage/routes/files/trash/+page.svelte
@@ -2,16 +2,26 @@
 	import { formatBytes } from '@axium/core/format';
 	import { forMime as iconForMime } from '@axium/core/icons';
 	import { FormDialog, Icon } from '@axium/server/components';
-	import { deleteItem } from '@axium/storage/client';
+	import { deleteItem, updateItemMetadata } from '@axium/storage/client';
 	import '@axium/storage/styles/list';
 	import type { PageProps } from './$types';
 
 	const { data }: PageProps = $props();
 	let items = $state(data.items);
-	let dialog = $state<HTMLDialogElement>();
+	let restoreDialog = $state<HTMLDialogElement>()!;
+	let deleteDialog = $state<HTMLDialogElement>()!;
 
 	let activeIndex = $state<number>(-1);
-	const activeItem = $derived(items[activeIndex]);
+	const activeItem = $derived(activeIndex == -1 ? null : items[activeIndex]);
+
+	function action(index: number, dialog: () => HTMLDialogElement) {
+		return (e: Event) => {
+			e.stopPropagation();
+			e.preventDefault();
+			activeIndex = index;
+			dialog().showModal();
+		};
+	}
 </script>
 
 <svelte:head>
@@ -31,16 +41,11 @@
 			<span class="name">{item.name}</span>
 			<span>{item.modifiedAt.toLocaleString()}</span>
 			<span>{formatBytes(item.size)}</span>
-			<span
-				class="action"
-				onclick={(e: Event) => {
-					e.stopPropagation();
-					e.preventDefault();
-					activeIndex = i;
-					dialog?.showModal();
-				}}
-			>
-				<Icon i="trash" --size="14px" --fill="#c44" />
+			<span class="action" onclick={action(i, () => restoreDialog)}>
+				<Icon i="rotate-left" --size="14px" />
+			</span>
+			<span class="action" onclick={action(i, () => deleteDialog)}>
+				<Icon i="trash-can-xmark" --size="14px" --fill="#c44" />
 			</span>
 		</div>
 	{:else}
@@ -48,25 +53,40 @@
 	{/each}
 </div>
 
+{#snippet _name()}
+	{#if activeItem?.name}<strong>{activeItem.name.length > 23 ? activeItem.name.slice(0, 20) + '...' : activeItem.name}</strong>
+	{:else}this
+	{/if}
+{/snippet}
+
 <FormDialog
-	bind:dialog
+	bind:dialog={restoreDialog}
+	submitText="Restore"
+	submit={async () => {
+		if (!activeItem) throw 'No item is selected';
+		await updateItemMetadata(activeItem.id, { trash: false });
+		items.splice(activeIndex, 1);
+	}}
+>
+	<p>Restore {@render _name()}?</p>
+</FormDialog>
+<FormDialog
+	bind:dialog={deleteDialog}
 	submitText="Delete"
 	submitDanger
 	submit={async () => {
+		if (!activeItem) throw 'No item is selected';
 		await deleteItem(activeItem.id);
-		if (activeIndex != -1) items.splice(activeIndex, 1);
+		items.splice(activeIndex, 1);
 	}}
 >
 	<p>
-		Are you sure you want to permanently delete
-		{#if activeItem?.name}<strong>{activeItem.name.length > 23 ? activeItem.name.slice(0, 20) + '...' : activeItem.name}</strong>
-		{:else}this
-		{/if}?
+		Are you sure you want to permanently delete {@render _name()}?
 	</p>
 </FormDialog>
 
 <style>
 	.list-item {
-		grid-template-columns: 1em 4fr 15em 5em 1em !important;
+		grid-template-columns: 1em 4fr 15em 5em 1em 1em;
 	}
 </style>
